Update to 0eb40ad2cd45f7b02b138b1a4090966905ed0fec

+++ resources/app/extensions/chrome-debug/out/chromeDebugAdapter.js
@@ -1,15 +1,18 @@
+"use strict";
 /*---------------------------------------------------------
  * Copyright (C) Microsoft Corporation. All rights reserved.
  *--------------------------------------------------------*/
-"use strict";
-const vscode_chrome_debug_core_1 = require('vscode-chrome-debug-core');
-const child_process_1 = require('child_process');
-const utils = require('./utils');
-const path = require('path');
+Object.defineProperty(exports, "__esModule", { value: true });
+const path = require("path");
+const vscode_chrome_debug_core_1 = require("vscode-chrome-debug-core");
+const child_process_1 = require("child_process");
+const utils = require("./utils");
 const DefaultWebSourceMapPathOverrides = {
+    'webpack:///./~/*': '${webRoot}/node_modules/*',
     'webpack:///./*': '${webRoot}/*',
     'webpack:///*': '*',
-    'meteor://💻app/*': '${webRoot}/*',
+    'webpack:///src/*': '${webRoot}/*',
+    'meteor://💻app/*': '${webRoot}/*'
 };
 class ChromeDebugAdapter extends vscode_chrome_debug_core_1.ChromeDebugAdapter {
     initialize(args) {
@@ -20,102 +23,58 @@ class ChromeDebugAdapter extends vscode_chrome_debug_core_1.ChromeDebugAdapter {
     }
     launch(args) {
         return super.launch(args).then(() => {
-            vscode_chrome_debug_core_1.logger.log('Using Kha from ' + args.kha + '\n', true);
-            let options = {
-                from: args.cwd,
-                to: path.join(args.cwd, 'build'),
-                projectfile: 'khafile.js',
-                target: 'debug-html5',
-                vr: 'none',
-                pch: false,
-                intermediate: '',
-                graphics: 'direct3d9',
-                visualstudio: 'vs2015',
-                kha: '',
-                haxe: '',
-                ogg: '',
-                aac: '',
-                mp3: '',
-                h264: '',
-                webm: '',
-                wmv: '',
-                theora: '',
-                kfx: '',
-                krafix: '',
-                ffmpeg: args.ffmpeg,
-                nokrafix: false,
-                embedflashassets: false,
-                compile: false,
-                run: false,
-                init: false,
-                name: 'Project',
-                server: false,
-                port: 8080,
-                debug: false,
-                silent: false,
-                watch: false
-            };
-            return require(path.join(args.kha, 'Tools/khamake/out/main.js')).run(options, {
-                info: message => {
-                    vscode_chrome_debug_core_1.logger.log(message, true);
-                }, error: message => {
-                    vscode_chrome_debug_core_1.logger.error(message, true);
-                }
-            }).then((value) => {
-                // Use vscode's electron
-                const chromePath = args.runtimeExecutable;
-                let chromeDir = chromePath;
-                if (chromePath.lastIndexOf('/') >= 0) {
-                    chromeDir = chromePath.substring(0, chromePath.lastIndexOf('/'));
-                }
-                else if (chromePath.lastIndexOf('\\') >= 0) {
-                    chromeDir = chromePath.substring(0, chromePath.lastIndexOf('\\'));
-                }
-                // Use custom electron
-                // const chromeDir = path.join(__dirname, '..', '..', '..', 'node_modules', 'electron', 'dist');
-                // let chromePath = chromeDir;
-                // if (process.platform === 'win32') chromePath = path.join(chromePath, 'electron.exe');
-                // else if (process.platform === 'darwin') chromePath = path.join(chromePath, 'Electron.app', 'Contents', 'MacOS', 'Electron');
-                // else chromePath = path.join(chromePath, 'electron');
-                // Start with remote debugging enabled
-                const port = args.port || Math.floor((Math.random() * 10000) + 10000);
-                const chromeArgs = ['--chromedebug', '--remote-debugging-port=' + port];
-                chromeArgs.push(path.resolve(args.cwd, args.file));
-                let launchUrl;
-                if (args.file) {
-                    launchUrl = vscode_chrome_debug_core_1.utils.pathToFileURL(path.join(args.cwd, args.file, 'index.html'));
-                }
-                else if (args.url) {
-                    launchUrl = args.url;
-                }
-                if (launchUrl) {
-                    chromeArgs.push(launchUrl);
-                }
-                vscode_chrome_debug_core_1.logger.log(`spawn('${chromePath}', ${JSON.stringify(chromeArgs)})`);
-                this._chromeProc = child_process_1.spawn(chromePath, chromeArgs, {
-                    detached: true,
-                    stdio: ['ignore'],
-                    cwd: chromeDir
-                });
-                this._chromeProc.unref();
-                this._chromeProc.on('error', (err) => {
-                    const errMsg = 'Chrome error: ' + err;
-                    vscode_chrome_debug_core_1.logger.error(errMsg);
-                    this.terminateSession(errMsg);
-                });
-                return this.doAttach(port, launchUrl, args.address);
-            }, (reason) => {
-                vscode_chrome_debug_core_1.logger.error('Launch canceled.', true);
-                return new Promise((resolve, reject) => {
-                    reject({ id: Math.floor(Math.random() * 100000), format: 'Compilation failed.' });
-                });
+            // Use vscode's electron
+            const chromePath = args.runtimeExecutable;
+            let chromeDir = chromePath;
+            if (chromePath.lastIndexOf('/') >= 0) {
+                chromeDir = chromePath.substring(0, chromePath.lastIndexOf('/'));
+            }
+            else if (chromePath.lastIndexOf('\\') >= 0) {
+                chromeDir = chromePath.substring(0, chromePath.lastIndexOf('\\'));
+            }
+            // Use custom electron
+            // const chromeDir = path.join(__dirname, '..', '..', '..', 'node_modules', 'electron', 'dist');
+            // let chromePath = chromeDir;
+            // if (process.platform === 'win32') chromePath = path.join(chromePath, 'electron.exe');
+            // else if (process.platform === 'darwin') chromePath = path.join(chromePath, 'Electron.app', 'Contents', 'MacOS', 'Electron');
+            // else chromePath = path.join(chromePath, 'electron');
+            // Start with remote debugging enabled
+            const port = args.port || Math.floor((Math.random() * 10000) + 10000);
+            const chromeArgs = ['--chromedebug', '--remote-debugging-port=' + port];
+            chromeArgs.push(path.resolve(args.cwd, args.file));
+            let launchUrl;
+            if (args.file) {
+                launchUrl = vscode_chrome_debug_core_1.utils.pathToFileURL(path.join(args.cwd, args.file, 'index.html'));
+            }
+            else if (args.url) {
+                launchUrl = args.url;
+            }
+            if (launchUrl) {
+                chromeArgs.push(launchUrl);
+            }
+            vscode_chrome_debug_core_1.logger.log(`spawn('${chromePath}', ${JSON.stringify(chromeArgs)})`);
+            this._chromeProc = this.spawnChrome(chromePath, chromeArgs, !!args.runtimeExecutable);
+            this._chromeProc.on('error', (err) => {
+                const errMsg = 'Chrome error: ' + err;
+                vscode_chrome_debug_core_1.logger.error(errMsg);
+                this.terminateSession(errMsg);
             });
+            return args.noDebug ? undefined :
+                this.doAttach(port, launchUrl || args.urlFilter, args.address, args.timeout);
         });
     }
     attach(args) {
+        if (args.urlFilter) {
+            args.url = args.urlFilter;
+        }
         return super.attach(args);
     }
     commonArgs(args) {
+        if (!args.webRoot && args.pathMapping && args.pathMapping['/']) {
+            // Adapt pathMapping['/'] as the webRoot when not set, since webRoot is explicitly used in many places
+            args.webRoot = args.pathMapping['/'];
+        }
+        args.sourceMaps = typeof args.sourceMaps === 'undefined' || args.sourceMaps;
         args.sourceMapPathOverrides = getSourceMapPathOverrides(args.webRoot, args.sourceMapPathOverrides);
         args.skipFileRegExps = ['^chrome-extension:.*'];
         super.commonArgs(args);
@@ -124,32 +83,93 @@ class ChromeDebugAdapter extends vscode_chrome_debug_core_1.ChromeDebugAdapter {
         return super.doAttach(port, targetUrl, address, timeout).then(() => {
             // Don't return this promise, a failure shouldn't fail attach
             this.globalEvaluate({ expression: 'navigator.userAgent', silent: true })
-                .then(evalResponse => vscode_chrome_debug_core_1.logger.log('Target userAgent: ' + evalResponse.result.value), err => vscode_chrome_debug_core_1.logger.log('Getting userAgent failed: ' + err.message));
+                .then(evalResponse => vscode_chrome_debug_core_1.logger.log('Target userAgent: ' + evalResponse.result.value), err => vscode_chrome_debug_core_1.logger.log('Getting userAgent failed: ' + err.message))
+                .then(() => {
+                const cacheDisabled = this._launchAttachArgs.disableNetworkCache || false;
+                this.chrome.Network.setCacheDisabled({ cacheDisabled });
+            });
         });
     }
     runConnection() {
-        return [...super.runConnection(), this.chrome.Page.enable()];
+        return [
+            ...super.runConnection(),
+            this.chrome.Page.enable(),
+            this.chrome.Network.enable({})
+        ];
     }
-    onPaused(notification) {
+    onPaused(notification, expectingStopReason) {
         this._overlayHelper.doAndCancel(() => this.chrome.Page.configureOverlay({ message: ChromeDebugAdapter.PAGE_PAUSE_MESSAGE }).catch(() => { }));
-        super.onPaused(notification);
+        super.onPaused(notification, expectingStopReason);
+    }
+    threadName() {
+        return 'Chrome';
     }
     onResumed() {
         this._overlayHelper.wait(() => this.chrome.Page.configureOverlay({}).catch(() => { }));
         super.onResumed();
     }
     disconnect() {
-        if (this._chromeProc) {
-            this._chromeProc.kill('SIGINT');
-            this._chromeProc = null;
+        const hadTerminated = this._hasTerminated;
+        // Disconnect before killing Chrome, because running "taskkill" when it's paused sometimes doesn't kill it
+        super.disconnect();
+        if (this._chromeProc && !hadTerminated) {
+            // Only kill Chrome if the 'disconnect' originated from vscode. If we previously terminated
+            // due to Chrome shutting down, or devtools taking over, don't kill Chrome.
+            if (vscode_chrome_debug_core_1.utils.getPlatform() === 0 /* Windows */ && this._chromePID) {
+                // Run synchronously because this process may be killed before exec() would run
+                const taskkillCmd = `taskkill /F /T /PID ${this._chromePID}`;
+                vscode_chrome_debug_core_1.logger.log(`Killing Chrome process by pid: ${taskkillCmd}`);
+                try {
+                    child_process_1.execSync(taskkillCmd);
+                }
+                catch (e) {
+                    // Can fail if Chrome was already open, and the process with _chromePID is gone.
+                    // Or if it already shut down for some reason.
+                }
+            }
+            else {
+                vscode_chrome_debug_core_1.logger.log('Killing Chrome process');
+                this._chromeProc.kill('SIGINT');
+            }
         }
-        return super.disconnect();
+        this._chromeProc = null;
     }
     /**
      * Opt-in event called when the 'reload' button in the debug widget is pressed
      */
     restart() {
         return this.chrome.Page.reload({ ignoreCache: true });
+    }
+    spawnChrome(chromePath, chromeArgs, usingRuntimeExecutable) {
+        if (vscode_chrome_debug_core_1.utils.getPlatform() === 0 /* Windows */ && !usingRuntimeExecutable) {
+            const chromeProc = child_process_1.fork(getChromeSpawnHelperPath(), [chromePath, ...chromeArgs], { execArgv: [], silent: true });
+            chromeProc.unref();
+            chromeProc.on('message', data => {
+                const pidStr = data.toString();
+                vscode_chrome_debug_core_1.logger.log('got chrome PID: ' + pidStr);
+                this._chromePID = parseInt(pidStr, 10);
+            });
+            chromeProc.on('error', (err) => {
+                const errMsg = 'chromeSpawnHelper error: ' + err;
+                vscode_chrome_debug_core_1.logger.error(errMsg);
+            });
+            chromeProc.stderr.on('data', data => {
+                vscode_chrome_debug_core_1.logger.error('[chromeSpawnHelper] ' + data.toString());
+            });
+            chromeProc.stdout.on('data', data => {
+                vscode_chrome_debug_core_1.logger.log('[chromeSpawnHelper] ' + data.toString());
+            });
+            return chromeProc;
+        }
+        else {
+            vscode_chrome_debug_core_1.logger.log(`spawn('${chromePath}', ${JSON.stringify(chromeArgs)})`);
+            const chromeProc = child_process_1.spawn(chromePath, chromeArgs, {
+                detached: true,
+                stdio: ['ignore'],
+            });
+            chromeProc.unref();
+            return chromeProc;
+        }
     }
 }
 ChromeDebugAdapter.PAGE_PAUSE_MESSAGE = 'Paused in Kode Studio';
@@ -181,5 +201,14 @@ function resolveWebRootPattern(webRoot, sourceMapPathOverrides, warnOnMissing) {
     }
     return resolvedOverrides;
 }
-exports.resolveWebRootPattern = resolveWebRootPattern;
-//# sourceMappingURL=https://ticino.blob.core.windows.net/sourcemaps/ebff2335d0f58a5b01ac50cb66737f4694ec73f3/extensions/chrome-debug/out/chromeDebugAdapter.js.map
+exports.resolveWebRootPattern = resolveWebRootPattern;
+function getChromeSpawnHelperPath() {
+    if (path.basename(__dirname) === 'src') {
+        // For tests
+        return path.join(__dirname, '../chromeSpawnHelper.js');
+    }
+    else {
+        return path.join(__dirname, 'chromeSpawnHelper.js');
+    }
+}
+//# sourceMappingURL=https://ticino.blob.core.windows.net/sourcemaps/0eb40ad2cd45f7b02b138b1a4090966905ed0fec/extensions/chrome-debug/out/chromeDebugAdapter.js.map
